functions/requirements.txt
firebase-functions
firebase-admin
google-genai
google-auth
httpx
pydantic
requests
python-dotenv
google-cloud-storage
google-cloud-firestore
